separate monitoring commands
