+++ odk2gn/gn2_utils.py
@@ -1,124 +1,28 @@
 import logging
 import os
 import csv
-import json
-import geojson
-from shapely.geometry import shape, Point, Polygon
+from shapely.geometry import shape
 from shapely.ops import transform
 from geoalchemy2.shape import from_shape
 
 import tempfile
-from sqlalchemy.orm.exc import NoResultFound
-from sqlalchemy.orm import relationship
-from sqlalchemy.sql import func, select, join
 from geonature.utils.env import DB
 from geonature.core.users.models import VUserslistForallMenu
-from geonature.core.gn_meta.models import TDatasets
 from geonature.core.gn_commons.models import TModules
-from geonature.core.gn_monitoring.models import TBaseSites
-from gn_module_monitoring.monitoring.models import (
-    TMonitoringModules,
-    TMonitoringSites,
-    TMonitoringSitesGroups,
-)
 
-from pypnnomenclature.models import TNomenclatures, BibNomenclaturesTypes, CorTaxrefNomenclature
 
-from odk2gn.monitoring_config import get_nomenclatures_fields
+from pypnnomenclature.models import TNomenclatures, CorTaxrefNomenclature
+
 from apptax.taxonomie.models import Taxref
 
 log = logging.getLogger("app")
 
-
-def get_monitoring_modules():
-    tab = []
-    modules = (TMonitoringModules).query.filter_by(type="monitoring_modules").all()
-    for module in modules:
-        tab.append(module)
-    return tab
 
 
 def get_module_code(id_module: int):
     module_code = (TModules.query.filter_by(id_module=id_module).one()).module_code
     return module_code
 
-
-def get_modules_info(module_code: str):
-    try:
-        module = TMonitoringModules.query.filter(
-            TMonitoringModules.module_code.ilike(module_code)
-        ).one()
-        return module
-    except NoResultFound:
-        log.error(f"No GeoNature module found for {module_code}")
-        raise
-
-
-def get_gn2_attachments_data(
-    module: TMonitoringModules,
-    skip_taxons: bool = False,
-    skip_observers: bool = False,
-    skip_jdd: bool = False,
-    skip_sites: bool = False,
-    skip_nomenclatures: bool = False,
-    skip_sites_groups: bool = False,
-):
-    files = {}
-    # Taxon
-    if not skip_taxons:
-        data = get_taxon_list(module.id_list_taxonomy)
-        files["gn_taxons.csv"] = to_csv(header=("cd_nom", "nom_complet", "nom_vern"), data=data)
-    # Observers
-    if not skip_observers:
-        data = get_observer_list(module.id_list_observer)
-        files["gn_observateurs.csv"] = to_csv(header=("id_role", "nom_complet"), data=data)
-    # JDD
-    if not skip_jdd:
-        data = format_jdd_list(module.datasets)
-        files["gn_jdds.csv"] = to_csv(header=("id_dataset", "dataset_name"), data=data)
-
-    # Sites
-    if not skip_sites:
-        data = get_site_list(module.id_module)
-        files["gn_sites.csv"] = to_csv(
-            header=("id_base_site", "base_site_name", "geometry"), data=data
-        )
-
-    if not skip_sites_groups:
-        data = get_site_groups_list(module.id_module)
-        files["gn_groupes.csv"] = to_csv(header=("id_sites_group", "sites_group_name"), data=data)
-
-    # Nomenclature
-    if not skip_nomenclatures:
-        n_fields = []
-        for niveau in ["site", "visit", "observation"]:
-            n_fields = n_fields + get_nomenclatures_fields(
-                module_code=module.module_code, niveau=niveau
-            )
-
-        nomenclatures = get_nomenclature_data(n_fields)
-        files["gn_nomenclatures.csv"] = to_csv(
-            header=("mnemonique", "id_nomenclature", "cd_nomenclature", "label_default"),
-            data=nomenclatures,
-        )
-
-    return files
-
-
-def get_site_groups_list(id_module: int):
-    """Return dict of TMonitoringSitesGroups
-
-    :param id_module: Identifier of the module
-    :type id_module : int"""
-
-    data = (
-        DB.session.query(TMonitoringSitesGroups)
-        .order_by(TMonitoringSitesGroups.sites_group_name)
-        .filter_by(id_module=id_module)
-        .all()
-    )
-
-    return [group.as_dict() for group in data]
 
 
 def get_taxon_list(id_liste: int):
@@ -142,31 +46,6 @@ def get_taxon_list(id_liste: int):
     return taxons
 
 
-def get_site_list(id_module: int):
-    """Return tuple of TBase site for module
-
-    :param id_module: Identifiant du module
-    :type id_module: int
-    """
-    data = (
-        DB.session.query(
-            TBaseSites.id_base_site,
-            TBaseSites.base_site_name,
-            func.concat(
-                func.st_y(func.st_centroid(TBaseSites.geom)),
-                " ",
-                func.st_x(func.st_centroid(TBaseSites.geom)),
-            ),
-        )
-        .order_by(TBaseSites.base_site_name)
-        .filter(TMonitoringSites.id_base_site == TBaseSites.id_base_site)
-        .filter(TMonitoringSites.id_module == id_module)
-        .all()
-    )
-    res = []
-    for d in data:
-        res.append({"id_base_site": d[0], "base_site_name": d[1], "geometry": d[2]})
-    return res
 
 
 def get_observer_list(id_liste: int):
@@ -295,6 +174,50 @@ def to_wkb(geojson):
     geom = transform(lambda x, y, z=None: (x, y), shape(geojson))
     return from_shape(geom, srid=4326)
 
+
+def get_and_post_medium(
+    project_id,
+    form_id,
+    uuid_sub,
+    filename,
+    monitoring_table,
+    media_type,
+    uuid_gn_object,
+):
+    # TODO : remove app context
+    img = get_attachment(project_id, form_id, uuid_sub, filename)
+    if img:
+        try:
+            uuid_sub = uuid_sub.split(":")[1]
+            medias_name = f"{uuid_sub}_{filename}"
+            table_location = (
+                DB.session.query(BibTablesLocation)
+                .filter_by(
+                    schema_name="gn_monitoring",
+                    table_name=monitoring_table,
+                )
+                .one()
+            )
+            media_type = (
+                DB.session.query(TNomenclatures)
+                .filter_by(mnemonique=media_type)
+                .filter(TNomenclatures.nomenclature_type.has(mnemonique="TYPE_MEDIA"))
+                .one()
+            )
+            media = {
+                "media_path": f"media/attachments/{medias_name}",
+                "uuid_attached_row": uuid_gn_object,
+                "id_table_location": table_location.id_table_location,
+                "id_nomenclature_media_type": media_type.id_nomenclature,
+            }
+
+            media = TMedias(**media)
+            DB.session.add(media)
+            DB.session.commit()
+            with open(BACKEND_DIR / "media" / "attachments" / medias_name, "wb") as out_file:
+                out_file.write(img.content)
+        except:
+            pass
 
 # def format_coords(geom):
 #     """removes the z coordinate of a geoJSON
