+++ odk2gn/monitoring/utils.py
@@ -5,6 +5,10 @@ import csv
 import json
 import datetime
 from shapely.geometry import Polygon, Point, LineString
+from sqlalchemy.sql import func, select, join
+
+
+from sqlalchemy.orm.exc import NoResultFound
 
 
 from gn_module_monitoring.monitoring.models import (
@@ -24,8 +28,11 @@ from gn_module_monitoring.monitoring.models import TMonitoringModules
 log = logging.getLogger("app")
 
 from pypnnomenclature.models import TNomenclatures, BibNomenclaturesTypes
-from odk2gn.gn2_utils import format_jdd_list, get_id_nomenclature_type_site, to_wkb
-
+from odk2gn.gn2_utils import (
+     format_jdd_list, get_id_nomenclature_type_site,
+     to_wkb, get_nomenclature_data, get_nomenclatures_fields,
+     get_observer_list, get_taxon_list, to_csv
+)
 
 def get_site_type_cd_nomenclature(monitoring_config):
     return monitoring_config["site"]["generic"]["id_nomenclature_type_site"]["value"][
@@ -247,3 +254,122 @@ def parse_and_create_obs(
             ].get("value")
     obs = TMonitoringObservations(**observation_dict_to_post)
     return obs
+
+
+def get_modules_info(module_code: str):
+    try:
+        module = TMonitoringModules.query.filter(
+            TMonitoringModules.module_code.ilike(module_code)
+        ).one()
+        return module
+    except NoResultFound:
+        log.error(f"No GeoNature module found for {module_code}")
+        raise
+
+
+def get_site_groups_list(id_module: int):
+    """Return dict of TMonitoringSitesGroups
+
+    :param id_module: Identifier of the module
+    :type id_module : int"""
+
+    data = (
+        DB.session.query(TMonitoringSitesGroups)
+        .order_by(TMonitoringSitesGroups.sites_group_name)
+        .filter_by(id_module=id_module)
+        .all()
+    )
+
+def get_gn2_attachments_data(
+    module: TMonitoringModules,
+    skip_taxons: bool = False,
+    skip_observers: bool = False,
+    skip_jdd: bool = False,
+    skip_sites: bool = False,
+    skip_nomenclatures: bool = False,
+    skip_sites_groups: bool = False,
+):
+    files = {}
+    # Taxon
+    if not skip_taxons:
+        data = get_taxon_list(module.id_list_taxonomy)
+        files["gn_taxons.csv"] = to_csv(header=("cd_nom", "nom_complet", "nom_vern"), data=data)
+    # Observers
+    if not skip_observers:
+        data = get_observer_list(module.id_list_observer)
+        files["gn_observateurs.csv"] = to_csv(header=("id_role", "nom_complet"), data=data)
+    # JDD
+    if not skip_jdd:
+        data = format_jdd_list(module.datasets)
+        files["gn_jdds.csv"] = to_csv(header=("id_dataset", "dataset_name"), data=data)
+
+    # Sites
+    if not skip_sites:
+        data = get_site_list(module.id_module)
+        files["gn_sites.csv"] = to_csv(
+            header=("id_base_site", "base_site_name", "geometry"), data=data
+        )
+
+    if not skip_sites_groups:
+        data = get_site_groups_list(module.id_module)
+        files["gn_groupes.csv"] = to_csv(header=("id_sites_group", "sites_group_name"), data=data)
+
+    # Nomenclature
+    if not skip_nomenclatures:
+        n_fields = []
+        for niveau in ["site", "visit", "observation"]:
+            n_fields = n_fields + get_nomenclatures_fields(
+                module_code=module.module_code, niveau=niveau
+            )
+
+        nomenclatures = get_nomenclature_data(n_fields)
+        files["gn_nomenclatures.csv"] = to_csv(
+            header=("mnemonique", "id_nomenclature", "cd_nomenclature", "label_default"),
+            data=nomenclatures,
+        )
+
+    return files
+
+
+def get_site_list(id_module: int):
+    """Return tuple of TBase site for module
+
+    :param id_module: Identifiant du module
+    :type id_module: int
+    """
+    data = (
+        DB.session.query(
+            TBaseSites.id_base_site,
+            TBaseSites.base_site_name,
+            func.concat(
+                func.st_y(func.st_centroid(TBaseSites.geom)),
+                " ",
+                func.st_x(func.st_centroid(TBaseSites.geom)),
+            ),
+        )
+        .order_by(TBaseSites.base_site_name)
+        .filter(TMonitoringSites.id_base_site == TBaseSites.id_base_site)
+        .filter(TMonitoringSites.id_module == id_module)
+        .all()
+    )
+    res = []
+    for d in data:
+        res.append({"id_base_site": d[0], "base_site_name": d[1], "geometry": d[2]})
+    return res
+
+
+
+def get_site_groups_list(id_module: int):
+    """Return dict of TMonitoringSitesGroups
+
+    :param id_module: Identifier of the module
+    :type id_module : int"""
+
+    data = (
+        DB.session.query(TMonitoringSitesGroups)
+        .order_by(TMonitoringSitesGroups.sites_group_name)
+        .filter_by(id_module=id_module)
+        .all()
+    )
+
+    return [group.as_dict() for group in data]
